[Fix] Refactor: Reorganize file structure for better modularity
- Moved home_screen.dart to modules/home
- Moved splash_screen.dart to modules
- Moved SearchField.dart to widgets/home
- Moved navigation_bar.dart to widgets
